Rename/refactor `GenericSensor` to `GenericInput`
