+++ src/io.rs
@@ -4,7 +4,6 @@ mod device;
 mod event;
 mod input;
 mod metadata;
-mod sensor;
 mod observer;
 mod command;
 mod output;
@@ -13,10 +12,9 @@ mod types;
 pub use calibrated::Calibrated;
 pub use device::*;
 pub use event::IOEvent;
-pub use input::Input;
+pub use input::*;
 pub use output::*;
 pub use metadata::DeviceMetadata;
-pub use sensor::*;
 pub use observer::*;
 pub use command::*;
 pub use types::*;
