+++ src/io/input.rs
@@ -1,10 +1,11 @@
 use crate::errors;
-use crate::helpers::Deferred;
-use crate::io::{Device, IOEvent, Publisher};
+use crate::helpers::{Deferrable, Deferred};
+use crate::io::{Device, DeviceMetadata, IdType, IODirection, IOEvent, IOKind, Publisher, SubscriberStrategy};
 use crate::io::types::{IdTraits, InputType, IOType};
-use crate::storage::{Container, Containerized};
+use crate::storage::{Container, Containerized, MappedCollection, OwnedLog};
 use chrono::{DateTime, Utc};
 use std::fmt::Formatter;
+use std::sync::{Arc, Mutex};
 
 /// Interface defining an input device
 /// It is used as a trait object and can be stored in a container using the `Containerized` trait.
@@ -30,7 +31,7 @@ pub trait Input: Device + Publisher {
     fn rx(&self) -> IOType;
 
     fn generate_event(&self, dt: DateTime<Utc>, value: Option<IOType>) -> IOEvent {
-        IOEvent::create(self, dt, value.unwrap_or_else(move || self.rx()))
+        IOEvent::generate(self, dt, value.unwrap_or_else(move || self.rx()))
     }
 
     fn read(&mut self, time: DateTime<Utc>) -> errors::Result<()>;
@@ -56,5 +57,77 @@ impl std::fmt::Debug for dyn Input {
             self.id(),
             self.metadata().kind
         )
+    }
+}
+
+#[derive(Default)]
+pub struct GenericInput {
+    metadata: DeviceMetadata,
+    pub log: Deferred<OwnedLog>,
+    subscribers: Vec<Deferred<Box<dyn SubscriberStrategy>>>,
+}
+
+
+impl Deferrable for GenericInput {
+    type Inner = InputType;
+    /// Return wrapped Sensor in
+    fn deferred(self) -> Deferred<Self::Inner> {
+        Arc::new(Mutex::new(Box::new(self)))
+    }
+}
+
+// Implement traits
+impl Device for GenericInput {
+    /// Creates a mock sensor which returns a value
+    ///
+    /// # Arguments
+    ///
+    /// * `name`: arbitrary name of sensor
+    /// * `id`: arbitrary, numeric ID to differentiate from other sensors
+    ///
+    /// returns: MockPhSensor
+    fn new(name: String, id: IdType, kind: Option<IOKind>, log: Deferred<OwnedLog>) -> Self where Self: Sized {
+        let kind = kind.unwrap_or_default();
+
+        let metadata: DeviceMetadata = DeviceMetadata::new(name, id, kind, IODirection::Input);
+        let subscribers = Vec::default();
+
+        GenericInput { metadata, log, subscribers }
+    }
+
+    fn metadata(&self) -> &DeviceMetadata {
+        &self.metadata
+    }
+}
+
+impl Input for GenericInput {
+    /// Return a mock value
+    fn rx(&self) -> IOType {
+        1.2
+    }
+
+    /// Get IOEvent, add to log, and propagate to subscribers
+    /// Primary interface method during polling.
+    fn read(&mut self, time: DateTime<Utc>) -> errors::Result<()> {
+        // get IOEvent
+        let event = self.generate_event(time, None);
+
+        // propagate to subscribers
+        self.notify(&event);
+
+        // add to log
+        let result = self.log.lock().unwrap().push(time, event);
+
+        result
+    }
+}
+
+impl Publisher for GenericInput {
+    fn subscribers(&mut self) -> &mut [Deferred<Box<dyn SubscriberStrategy>>] {
+        &mut self.subscribers
+    }
+
+    fn subscribe(&mut self, subscriber: Deferred<Box<dyn SubscriberStrategy>>) {
+        self.subscribers.push(subscriber)
     }
 }
