+++ src/storage/grouping.rs
@@ -8,15 +8,13 @@ use crate::io::InputContainer;
 use crate::settings::Settings;
 use crate::storage::{LogContainer, MappedCollection, Persistent};
 
-/// Mediator to periodically poll sensors of various types, and store the resulting `IOEvent` objects in a `Container`.
+/// Mediator to periodically poll input devices of various types, and store the resulting `IOEvent` objects in a `Container`.
 ///
-/// `poll()` is the primary callable and iterates through the `Sensor` container to call `get_event()` on each sensor.
+/// `poll()` is the primary callable and iterates through `InputContainers` to call `read()` on each input device.
 /// Resulting `IOEvent` objects are then added to the `log` container.
 ///
-/// The `interval` field indicates the duration between each poll and the `last_execution` field indicates the last time the poll method was executed
-///
-/// TODO: multithreaded polling. Implement `RwLock` or `Mutex` to synchronize access to the sensors and
-///       log containers in order to make the poll() function thread-safe.
+/// `interval` dictates the duration between each poll,
+/// and `last_execution` field is working memory to store the time of the last successful poll.
 pub struct PollGroup {
     name: String,
     last_execution: DateTime<Utc>,
@@ -38,8 +36,8 @@ impl PollGroup {
         let next_execution = self.last_execution + self.settings.interval;
 
         if next_execution <= Utc::now() {
-            for (_, sensor) in self.inputs.iter_mut() {
-                let result = sensor.lock().unwrap().read(next_execution);
+            for (_, input) in self.inputs.iter_mut() {
+                let result = input.lock().unwrap().read(next_execution);
                 results.push(result);
             }
             self.last_execution = next_execution;
@@ -55,7 +53,7 @@ impl PollGroup {
         let settings = settings.unwrap_or_else(|| Arc::new(Settings::default()));
         let last_execution = Utc::now() - settings.interval;
 
-        let sensors = <InputContainer<IdType>>::default();
+        let inputs = <InputContainer<IdType>>::default();
         let logs = Vec::default();
 
         Self {
@@ -63,7 +61,7 @@ impl PollGroup {
             settings,
             last_execution,
             logs,
-            inputs: sensors,
+            inputs,
         }
     }
 
@@ -98,7 +96,7 @@ impl PollGroup {
     /// Load each individual log
     /// # Notes
     /// This works because each log container should have it's own name upon initialization
-    /// from hardcoded sensors.
+    /// from hardcoded input devices.
     fn load_logs(&self, path: &Option<String>) -> Result<()> {
         let mut results = Vec::new();
         for log in self.logs.iter() {
@@ -111,7 +109,7 @@ impl PollGroup {
     /// Save each individual log
     /// # Notes
     /// This works because each log container should have it's own name upon initialization
-    /// from hardcoded sensors.
+    /// from hardcoded input devices.
     fn save_logs(&self, path: &Option<String>) -> Result<()> {
         let mut results = Vec::new();
         for log in self.logs.iter() {
